+++ src/libexpr/eval.cc
@@ -55,12 +55,6 @@ static void * allocBytes(size_t n)
 }
 
 
-void Bindings::sort()
-{
-    std::sort(begin(), end());
-}
-
-
 static void printValue(std::ostream & str, std::set<const Value *> & active, const Value & v)
 {
     if (active.find(&v) != active.end()) {
@@ -503,20 +497,6 @@ Env & EvalState::allocEnv(unsigned int size)
 }
 
 
-Value * EvalState::allocAttr(Value & vAttrs, const Symbol & name)
-{
-    Value * v = allocValue();
-    vAttrs.attrs->push_back(Attr(name, v));
-    return v;
-}
-
-
-Bindings * EvalState::allocBindings(Bindings::size_t capacity)
-{
-    return new (allocBytes(sizeof(Bindings) + sizeof(Attr) * capacity)) Bindings(capacity);
-}
-
-
 void EvalState::mkList(Value & v, unsigned int length)
 {
     clearValue(v);
@@ -524,16 +504,6 @@ void EvalState::mkList(Value & v, unsigned int length)
     v.list.length = length;
     v.list.elems = length ? (Value * *) allocBytes(length * sizeof(Value *)) : 0;
     nrListElems += length;
-}
-
-
-void EvalState::mkAttrs(Value & v, unsigned int expected)
-{
-    clearValue(v);
-    v.type = tAttrs;
-    v.attrs = allocBindings(expected);
-    nrAttrsets++;
-    nrAttrsInAttrsets += expected;
 }
